* dev: (95 commits)
  update ahj db attachment/link/contact type ids for UAT
  show int fields in tournament custom fields
  fix data_type_id returned for data view fields
  allow deleting docs from ahj databases
  new endpoint for mobile
  fix tourney query
  fix a couple of queries
  work queue clarity improvements
  function to copy ahj value to other cfgas
  change layout of work queue history
  data view fields used as requirements :fingers_crossed:
  moar scott columns
  add column for scott
  Remove stuff that I shouldn't have committed
  add commission_forfeited_by_closer field
  copy another thing on duplicate
  copy more info when a user duplicates an action in admin
  turn off local br cron by default
  fix insert on user residual
  Adding 'last 30 days' company dashboard option
  ...

+++ client/src/views/flow/components/AttachmentsTable.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <v-dialog persistent :width="1000" v-model="showCoversheetModal"
+    <v-dialog persistent  v-model="showCoversheetModal"
               content-class="coversheet-modal-content">
       <AttachmentCoversheetModal :existing-attachment="selectedFile"
                                  :show-modal="showCoversheetModal"
@@ -67,17 +67,16 @@
           <v-btn icon text color="primary" class="button-position" :href="item.presignedUrl" @mouseover="buttonHovered = true" @mouseleave="buttonHovered = false">
             <v-icon size="25" >mdi-tray-arrow-down</v-icon>
           </v-btn>
-          <ConfirmationDialog
-            :open-dialog="attachmentDeleteConfirm"
-            :retain-focus="false"
-            @confirm="deleteAttachment"
-            @close-dialog="closeDeleteDialog"
-
-          >Are you sure you want to delete {{ attachmentToDeleteName }}?
-          </ConfirmationDialog>
         </v-col>
       </v-row>
     </v-container>
+    <ConfirmationDialog
+        :open-dialog="attachmentDeleteConfirm"
+        :retain-focus="false"
+        @confirm="deleteAttachment"
+        @close-dialog="closeDeleteDialog">
+      Are you sure you want to delete <b>{{ attachmentToDeleteName }}</b>?
+    </ConfirmationDialog>
   </div>
 </template>
 
@@ -183,7 +182,7 @@ export default {
       const id = this.attachmentToDelete.id
       try {
         await deleteAttachment(id)
-        this.deleteCallback(id)
+        if(this.deleteCallback) {this.deleteCallback(id)}
         //this value tells the right pane to update when a file is deleted
         this.$store.commit(ProjectMutations.INCREMENT_RELOAD_KEY)
